+++ sentry-rails/spec/sentry/rails/activejob_spec.rb
@@ -27,6 +27,12 @@ class FailedWithExtraJob < FailedJob
   end
 end
 
+class JobWithArgument < ActiveJob::Base
+  def perform(*args, integer:, post:, **options)
+    raise "foo"
+  end
+end
+
 class QueryPostJob < ActiveJob::Base
   self.logger = nil
 
@@ -76,6 +82,52 @@ RSpec.describe "ActiveJob integration" do
     expect(NormalJob.perform_now).to eq("foo")
   end
 
+  describe "ActiveJob arguments serialization" do
+    it "serializes ActiveRecord arguments in globalid form" do
+      post = Post.create!
+      post2 = Post.create!
+
+      expect do
+        JobWithArgument.perform_now("foo", { bar: Sentry }, integer: 1, post: post, nested: { another_level: { post: post2 } })
+      end.to raise_error(RuntimeError)
+
+      event = transport.events.last.to_json_compatible
+      expect(event.dig("extra", "arguments")).to eq(
+        [
+          "foo",
+          { "bar" => "Sentry" },
+          {
+            "integer" => 1,
+            "post" => "gid://rails-test-app/Post/#{post.id}",
+            "nested" => { "another_level" => { "post" => "gid://rails-test-app/Post/#{post2.id}" } }
+          }
+        ]
+      )
+    end
+
+    it "handles problematic globalid conversion gracefully" do
+      post = Post.create!
+
+      def post.to_global_id
+        raise
+      end
+
+      expect do
+        JobWithArgument.perform_now(integer: 1, post: post)
+      end.to raise_error(RuntimeError)
+
+      event = transport.events.last.to_json_compatible
+      expect(event.dig("extra", "arguments")).to eq(
+        [
+          {
+            "integer" => 1,
+            "post" => post.to_s,
+          }
+        ]
+      )
+    end
+  end
+
   it "adds useful context to extra" do
     expect { FailedJob.perform_now }.to raise_error(FailedJob::TestError)
 
@@ -119,7 +171,7 @@ RSpec.describe "ActiveJob integration" do
       expect(transaction.contexts.dig(:trace, :status)).to eq("ok")
 
       expect(transaction.spans.count).to eq(1)
-      expect(transaction.spans.first[:op]).to eq("sql.active_record")
+      expect(transaction.spans.first[:op]).to eq("db.sql.active_record")
     end
 
     context "with error" do